+++ src/Application/Interface.ts
@@ -5,7 +5,6 @@
  * @library Application/Interface
  * @includes IConfig Application/_Interface/IConfig
  * @includes IEnv Application/_Interface/IEnv
- * @includes IHead Application/_Interface/IHead
  * @includes IHttpRequest Application/_Interface/IHttpRequest
  * @includes IHttpResponse Application/_Interface/IHttpResponse
  * @public
@@ -21,7 +20,7 @@ export * from 'Application/_Interface/IEnv';
 // export * from 'Application/_Interface/IStore';
 // export * from 'Application/_Interface/IStateReceiver';
 // export * from 'Application/_Interface/IRequest';
-export * from 'Application/_Interface/IHead';
+// export * from 'Application/_Interface/IHead';
 export * from 'Application/_Interface/IBody';
 export * from 'Application/_Interface/IHttpRequest';
 export * from 'Application/_Interface/IHttpResponse';
