+++ src/Application/_Page/_head/Factory.ts
@@ -1,9 +1,9 @@
 ///// <amd-module name="Application/_Page/_head/Factory" />
 
-import { IHeadTagAttrs, IHeadTagEventHandlers } from 'Application/_Interface/IHead';
 import BaseElement, { IHeadElement, IHeadElementAspect } from "Application/_Page/_head/BaseElement";
 import Element from "Application/_Page/_head/Element";
 import ElementPS from "Application/_Page/_head/ElementPS";
+import type { IHeadTagAttrs, IHeadTagEventHandlers } from 'Application/_Page/_head/IHead';
 import TitleAspect from "Application/_Page/_head/Title";
 import ViewPortAspect from "Application/_Page/_head/ViewPort";
