+++ src/Application/_Page/_head/ViewPort.ts
@@ -1,7 +1,7 @@
 ///// <amd-module name="Application/_Page/_head/BaseElement" />
 
-import { IHeadTag, JML } from 'Application/_Interface/IHead';
 import BaseElement, { IHeadElementAspect } from 'Application/_Page/_head/BaseElement';
+import type { IHeadTag, JML } from 'Application/_Page/_head/IHead';
 
 /**
  * Аспект для уникального элемента типа meta с параметрами для viewport
