+++ src/Application/_Page/_jslinks/JSLinksElement.ts
@@ -1,7 +1,7 @@
 ///// <amd-module name="Application/_Page/_jslinks/JSLinksElement" />
 
-import { IHeadTagAttrs, IHeadTagEventHandlers } from 'Application/_Interface/IHead';
 import ElementPS from 'Application/_Page/_head/ElementPS';
+import type { IHeadTagAttrs, IHeadTagEventHandlers } from 'Application/_Page/_head/IHead';
 
 /**
  * Элемент класса jslinks, отнаследованный от ElementPS
