+++ test/AppUnit/Page/Head/Element.test.ts
@@ -1,6 +1,6 @@
 import { assert } from 'chai';
 import { create } from 'Application/_Page/_head/Factory';
-import { IHeadTagAttrs } from 'Application/Interface';
+import type { IHeadTagAttrs } from 'Application/Page';
 
 const ALL_KEYS_ELEMENT = ['_name', '_attrs', '_content', '_eventHandlers'];
 const KEY_TITLE = ['_content'];
